HSEARCH-3544 Merge SearchQueryContextImpl and SearchQueryResultContextImpl
We don't need those to be separate classes, we just need the interfaces
they implement to be clearly separate to force users to set a predicate.

Having a single implementing class will make it easier to introduce
extensions.
